requirements.txt
colorama>=0.4.6
dnspython>=2.4.0
